+++ lib/polishgeeks/dev-tools/logger.rb
@@ -6,7 +6,7 @@ module PolishGeeks
       CLASS_NAME_REGEXP = /^.*::/
 
       # Method will print appropriate output with colors and comments
-      # @param [PolishGeeks::DevTools::Command::Base] task that was executed
+      # @param [PolishGeeks::DevTools::Commands::Base] task that was executed
       # @raise [PolishGeeks::DevTools::Logger::RequirementsError] raised when task
       #   has failed
       # @example Log output of a current_task
@@ -24,7 +24,7 @@ module PolishGeeks
 
       # Will print message when task didn't fail while running
       # @note Handles also generators that never fail
-      # @param [PolishGeeks::DevTools::Command::Base] task that was executed
+      # @param [PolishGeeks::DevTools::Commands::Base] task that was executed
       # @raise [PolishGeeks::DevTools::Errors::UnknownTaskType] raised when task type is
       #   not supported (we don't know how to handle it)
       def info(task)
@@ -37,7 +37,7 @@ module PolishGeeks
       end
 
       # Will print message when task did fail while running
-      # @param [PolishGeeks::DevTools::Command::Base] task that was executed
+      # @param [PolishGeeks::DevTools::Commands::Base] task that was executed
       def fatal(task)
         printf('%-30s %20s', "#{label(task)} ", "\e[31mFAILURE\e[0m\n\n")
         puts task.error_message + "\n"
@@ -45,7 +45,7 @@ module PolishGeeks
 
       # Generates a label describing a given task
       # @note Will use a task class name if task doesn't respond to label method
-      # @param [PolishGeeks::DevTools::Command::Base] task that was executed
+      # @param [PolishGeeks::DevTools::Commands::Base] task that was executed
       # @return [String] label describing task
       # @example Get label for a current_task task
       #   label(current_task) #=> 'Rubocop'