+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.1
 - Extracted all errors to PolishGeeks::DevTools::Errors namespace
 - Add support for validators
+- Rename Command to Commands namespace
 
 ## 1.2.0
 - Added EmptyMethod command which checks if some files have empty methods